HBASE-26483. openFile() and createFile()
adds a hadoop-3.3 source tree, where the production source implements the new methods
in a subclass of HBaseObjectStoreSemantics called HBase33ObjectStoreSemantics.
(obviously, applications will need to bind to this for the semantics)

Maven only adds the new source to the build with the hadoop3.3 profile.

How the existing tests bind to the store is made configurable in XML, and
will switch on hadoop 3.3 test runs to the new class throughout the contract
tests. This ensures that tests of the relevant APIs through the contract tests
are done.

There's a new test TestCreateFile to test the api and verify that the streams
returned by the builders are locking and exclusive.

This test is failing. Locking streams are coming back, but there's no
conflict between the two streams.

Change-Id: I2a4ba8116aee293e6fccac752143511cf0eb8883