HBASE-26483. binding to the extended hboss client
The reason the createFile test was failing is that and I was still binding to the parent version of hboss, the one which doesn't support the new APIs.

This commit tries to set things up properly and to use StreamCapabilities to actually ask if the FS supports the new APIs.

I'm going to switch to pathcapabilities there even though as that is a 3.3.3+ API only, I cannot add a probe for HBoss semantics in the base class.

I'm concluding that it is not only very convoluted to get all this to compile and test, the risk of deploying the 3.2 compatible client yet invoking those 3.3 APIs is quite high. If I ended up doing it in tests, I it will inevitably happen in production, which is very dangerous.

Things would work a lot easier if HBoss moved to being hadoop 3.3.0+ only.

Change-Id: Id2c8e0ed95b92df354d2e23956a66e1150e3cc01